pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pipelines"
version = "0.1.0"
description = "Data pipelines for Steam data sources"
requires-python = ">=3.10"

dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
airflow_dockerfile = [
    "apache-airflow>=2.10.3",
    "apache-airflow-providers-apache-spark>=4.0.0",
    "pyspark>=3.5.3",
    "minio>=7.2.0",
    "clickhouse-driver>=0.2.9",
]
spark_dockerfile = [
    "clickhouse-driver>=0.2.9",
]
dev = ["pytest>=7.0.0", "minio>=7.2.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/pipelines"]
